refactor: unify map and greening sols to one page
- merged map+page and green_solutions into a consolidated explore page
- redesigned page logic to show map first, using interactions to reveal
  data
- implement floating selection pill for POI/barangay toggling
- relocate map options to top-right to avoid mobile overlap
- responsive pattern for greening solutions popup (sidebar for desktop;
  bottom sheet for mobile)
- fixed mobile usability issues (buggy UI)
- updated navbar and landing page prefetching to reflect new structure
