+++ src/app/explore/barangaydetails.tsx
@@ -25,10 +25,10 @@ export default function BarangayMetricItem({
   return (
     <div
       className="
-        w-full flex flex-col items-center justify-between
-        bg-white/80 backdrop-blur-md gap-3
+        w-full flex items-center gap-3
+        bg-white/80 backdrop-blur-md
         border border-neutral-200
-        rounded-2xl p-5
+        rounded-2xl p-3
         transition-all duration-300
         hover:shadow-xl hover:shadow-neutral-200/40
         hover:border-primary-green/30
@@ -36,32 +36,32 @@ export default function BarangayMetricItem({
       "
     >
       <div
-        className={`p-4 flex items-center justify-center rounded-2xl 
-          ${bgColor} bg-opacity-10 group-hover:bg-opacity-20 transition-all duration-300`}
+        className={`p-2.5 flex items-center justify-center rounded-xl 
+          ${bgColor} bg-opacity-10 group-hover:bg-opacity-20 transition-all duration-300 shrink-0`}
       >
         <Icon
-          size={32}
+          size={20}
           className={`${textColor} group-hover:scale-110 transition-transform duration-300`}
         />
       </div>
 
-      <div className="flex flex-col items-center gap-1">
-        <span className="text-[10px] font-bold text-neutral-400 uppercase tracking-widest text-center">
+      <div className="flex flex-col items-start gap-0 min-w-0">
+        <span className="text-[9px] font-black text-neutral-400 uppercase tracking-widest truncate w-full">
           {label}
         </span>
 
         <div className="flex items-baseline gap-0.5">
           {isTemperature ? (
-            <span className={`font-bold text-2xl font-poppins ${textColor}`}>
+            <span className={`font-black text-base font-poppins ${textColor}`}>
               {value?.toFixed(0)}
             </span>
           ) : (
-            <span className={`font-bold text-2xl font-poppins ${textColor}`}>
+            <span className={`font-black text-base font-poppins ${textColor}`}>
               {value?.toFixed(2)}
             </span>
           )}
           {isTemperature && (
-            <span className={`text-sm font-bold ${textColor}`}>°C</span>
+            <span className={`text-[9px] font-black ${textColor}`}>°C</span>
           )}
         </div>
       </div>
